+++ packages/archives/src/contact/server/services/bootstrap.ts
@@ -1,5 +1,5 @@
 import { ZodSchemaValidator } from "@otuekong-portfolio/common";
-import { ConfiguredResendClient } from "@otuekong-portfolio/infrastructure-server";
+import { ConfiguredResendClient, EnvironmentRegistry } from "@otuekong-portfolio/infrastructure-server";
 
 import { ContactFormSchema } from "../../core";
 
@@ -8,5 +8,5 @@ import SendEmailHandler from "./sendEmailHandler";
 export const ConfiguredSendEmailHandler = new SendEmailHandler(
 	ConfiguredResendClient,
 	new ZodSchemaValidator(ContactFormSchema),
-	"contact@otuekongbassey.com"
+	EnvironmentRegistry
 );