+++ packages/curio/CHANGELOG.md
@@ -29,3 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix `useIsTruncated` to ensure the state only updates when the truncation status actually flips, and it uses `useLayoutEffect` to decouple logic from the ref assignment.
 * Fix `Typography` to memo-ize bits to prevent unnecessary changes from `useIsTruncated` and to stablize the props to createElement
 * Fix `Tooltip` to only display when there's a content and always display a span to prevent `Typography` from unmounting and remounting.
+
+### Removed
+* Remove the unused `phone` icon